requirements.txt
Flask
Flask-Cors
python-dotenv
google-generativeai
langchain-community
langchain-google-genai
pypdf2
gunicorn
